moved custom splitter to its own file
